+++ modules/sources/contracts/index.ts
@@ -1,5 +1,5 @@
 export * from './fetch-erc20-headers';
-export * from './v3/fetch-weighted-pools-data';
+export * from './v3/fetch-pool-type-data';
 export * from './v3/fetch-pool-data';
 export * from './v3/vault-client';
 export * from './v3/pools-client';