+++ prisma/migrations/20241107094621_remove_desc_from_hooks/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `description` on the `PrismaHookReviewData` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE "PrismaHookReviewData" DROP COLUMN "description";