+++ schema.ts
@@ -28,6 +28,17 @@ export interface Scalars {
     JSON: any;
 }
 
+/** The review data for the ERC4626 token */
+export interface Erc4626ReviewData {
+    __typename?: 'Erc4626ReviewData';
+    /** The filename of the review of the ERC4626 */
+    reviewFile: Scalars['String'];
+    /** A summary of the ERC4626 review, usually just says safe or unsafe */
+    summary: Scalars['String'];
+    /** Warnings associated with the ERC4626 */
+    warnings: Array<Scalars['String']>;
+}
+
 export interface GqlBalancePoolAprItem {
     __typename?: 'GqlBalancePoolAprItem';
     apr: GqlPoolAprValue;
@@ -134,7 +145,7 @@ export interface GqlNestedPool {
     __typename?: 'GqlNestedPool';
     /** Address of the pool. */
     address: Scalars['Bytes'];
-    /** Price rate of the Balancer Pool Token (BPT). */
+    /** Price rate of this pool or the Balancer Pool Token (BPT). */
     bptPriceRate: Scalars['BigDecimal'];
     /** Timestamp of when the pool was created. */
     createTime: Scalars['Int'];
@@ -363,10 +374,7 @@ export type GqlPoolAprValue = GqlPoolAprRange | GqlPoolAprTotal;
 export interface GqlPoolBase {
     /** The contract address of the pool. */
     address: Scalars['Bytes'];
-    /**
-     * Returns all pool tokens, including any nested tokens and phantom BPTs on one level.
-     * @deprecated Use poolTokens instead
-     */
+    /** Returns all pool tokens, including any nested tokens and phantom BPTs on one level. */
     allTokens: Array<GqlPoolTokenExpanded>;
     /** List of categories assigned by the team based on external factors */
     categories?: Maybe<Array<Maybe<GqlPoolFilterCategory>>>;
@@ -404,7 +412,7 @@ export interface GqlPoolBase {
     name: Scalars['String'];
     /** The wallet address of the owner of the pool. Pool owners can set certain properties like swapFees or AMP. */
     owner?: Maybe<Scalars['Bytes']>;
-    /** Returns all pool tokens, including BPTs and nested pools if there are any. Only one nested level deep. */
+    /** Returns pool tokens, including BPTs and nested pools and their pool tokens if there are any. Only one nested level deep. */
     poolTokens: Array<GqlPoolTokenDetail>;
     /** The protocol version on which the pool is deployed, 1, 2 or 3 */
     protocolVersion: Scalars['Int'];
@@ -1070,6 +1078,7 @@ export interface GqlPoolStable extends GqlPoolBase {
     address: Scalars['Bytes'];
     allTokens: Array<GqlPoolTokenExpanded>;
     amp: Scalars['BigInt'];
+    bptPriceRate: Scalars['BigDecimal'];
     categories?: Maybe<Array<Maybe<GqlPoolFilterCategory>>>;
     chain: GqlChain;
     createTime: Scalars['Int'];
@@ -1355,6 +1364,8 @@ export interface GqlPoolTokenDetail {
     balanceUSD: Scalars['BigDecimal'];
     /** Decimals of the pool token. */
     decimals: Scalars['Int'];
+    /** The ERC4626 review data for the token */
+    erc4626ReviewData?: Maybe<Erc4626ReviewData>;
     /** Indicates whether this token is a BPT and therefor has a nested pool. */
     hasNestedPool: Scalars['Boolean'];
     /** Id of the token. A combination of pool id and token address. */
@@ -1379,7 +1390,7 @@ export interface GqlPoolTokenDetail {
     scalingFactor?: Maybe<Scalars['BigDecimal']>;
     /** Symbol of the pool token. */
     symbol: Scalars['String'];
-    /** If it is an Erc4262, this will be the underlying token if present in the API. */
+    /** If it is an ERC4626, this will be the underlying token if present in the API. */
     underlyingToken?: Maybe<GqlToken>;
     /** The weight of the token in the pool if it is a weighted pool, null otherwise */
     weight?: Maybe<Scalars['BigDecimal']>;
@@ -1896,6 +1907,8 @@ export interface GqlToken {
     description?: Maybe<Scalars['String']>;
     /** The Discord URL of the token */
     discordUrl?: Maybe<Scalars['String']>;
+    /** The ERC4626 review data for the token */
+    erc4626ReviewData?: Maybe<Erc4626ReviewData>;
     /** Whether the token is considered an ERC4626 token. */
     isErc4626: Scalars['Boolean'];
     /** The logo URI of the token */
@@ -1919,6 +1932,8 @@ export interface GqlToken {
     tradable: Scalars['Boolean'];
     /** The Twitter username of the token */
     twitterUsername?: Maybe<Scalars['String']>;
+    /** The ERC4626 underlying token address, if applicable. */
+    underlyingTokenAddress?: Maybe<Scalars['String']>;
     /** The website URL of the token */
     websiteUrl?: Maybe<Scalars['String']>;
 }
@@ -1984,6 +1999,12 @@ export interface GqlTokenDynamicData {
     tokenAddress: Scalars['String'];
     /** The timestamp when the data was last updated */
     updatedAt: Scalars['String'];
+}
+
+/** Provide filters for tokens */
+export interface GqlTokenFilter {
+    /** Only return tokens with these addresses */
+    tokensIn?: InputMaybe<Array<Scalars['String']>>;
 }
 
 /** Result of the poolReloadPools mutation */
@@ -2140,6 +2161,8 @@ export interface Hook {
     enableHookAdjustedAmounts: Scalars['Boolean'];
     /** List of pools using the hook */
     poolsIds?: Maybe<Array<Maybe<Scalars['String']>>>;
+    /** The review for this hook if applicable. */
+    reviewData?: Maybe<HookReviewData>;
     shouldCallAfterAddLiquidity: Scalars['Boolean'];
     shouldCallAfterInitialize: Scalars['Boolean'];
     shouldCallAfterRemoveLiquidity: Scalars['Boolean'];
@@ -2157,6 +2180,19 @@ export interface HookData {
     addLiquidityFeePercentage?: Maybe<Scalars['String']>;
     removeLiquidityFeePercentage?: Maybe<Scalars['String']>;
     swapFeePercentage?: Maybe<Scalars['String']>;
+}
+
+/** Represents the review data for the hook */
+export interface HookReviewData {
+    __typename?: 'HookReviewData';
+    /** The name of the hook */
+    name: Scalars['String'];
+    /** The filename of the review of the hook */
+    reviewFile: Scalars['String'];
+    /** A summary of the hook review, usually just says safe or unsafe */
+    summary: Scalars['String'];
+    /** Warnings associated with the hook */
+    warnings: Array<Scalars['String']>;
 }
 
 /** Liquidity management settings for v3 pools. */
@@ -2553,6 +2589,7 @@ export interface QueryTokenGetTokenDynamicDataArgs {
 
 export interface QueryTokenGetTokensArgs {
     chains?: InputMaybe<Array<GqlChain>>;
+    where?: InputMaybe<GqlTokenFilter>;
 }
 
 export interface QueryTokenGetTokensDataArgs {
@@ -2704,6 +2741,7 @@ export type ResolversTypes = ResolversObject<{
     Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
     Bytes: ResolverTypeWrapper<Scalars['Bytes']>;
     Date: ResolverTypeWrapper<Scalars['Date']>;
+    Erc4626ReviewData: ResolverTypeWrapper<Erc4626ReviewData>;
     Float: ResolverTypeWrapper<Scalars['Float']>;
     GqlBalancePoolAprItem: ResolverTypeWrapper<
         Omit<GqlBalancePoolAprItem, 'apr'> & { apr: ResolversTypes['GqlPoolAprValue'] }
@@ -2869,6 +2907,7 @@ export type ResolversTypes = ResolversObject<{
     GqlTokenChartDataRange: GqlTokenChartDataRange;
     GqlTokenData: ResolverTypeWrapper<GqlTokenData>;
     GqlTokenDynamicData: ResolverTypeWrapper<GqlTokenDynamicData>;
+    GqlTokenFilter: GqlTokenFilter;
     GqlTokenMutationResult: ResolverTypeWrapper<GqlTokenMutationResult>;
     GqlTokenPrice: ResolverTypeWrapper<GqlTokenPrice>;
     GqlTokenPriceChartDataItem: ResolverTypeWrapper<GqlTokenPriceChartDataItem>;
@@ -2885,6 +2924,7 @@ export type ResolversTypes = ResolversObject<{
     GqlVotingPool: ResolverTypeWrapper<GqlVotingPool>;
     Hook: ResolverTypeWrapper<Hook>;
     HookData: ResolverTypeWrapper<HookData>;
+    HookReviewData: ResolverTypeWrapper<HookReviewData>;
     ID: ResolverTypeWrapper<Scalars['ID']>;
     Int: ResolverTypeWrapper<Scalars['Int']>;
     JSON: ResolverTypeWrapper<Scalars['JSON']>;
@@ -2905,6 +2945,7 @@ export type ResolversParentTypes = ResolversObject<{
     Boolean: Scalars['Boolean'];
     Bytes: Scalars['Bytes'];
     Date: Scalars['Date'];
+    Erc4626ReviewData: Erc4626ReviewData;
     Float: Scalars['Float'];
     GqlBalancePoolAprItem: Omit<GqlBalancePoolAprItem, 'apr'> & { apr: ResolversParentTypes['GqlPoolAprValue'] };
     GqlBalancePoolAprSubItem: Omit<GqlBalancePoolAprSubItem, 'apr'> & { apr: ResolversParentTypes['GqlPoolAprValue'] };
@@ -3041,6 +3082,7 @@ export type ResolversParentTypes = ResolversObject<{
     GqlTokenCandlestickChartDataItem: GqlTokenCandlestickChartDataItem;
     GqlTokenData: GqlTokenData;
     GqlTokenDynamicData: GqlTokenDynamicData;
+    GqlTokenFilter: GqlTokenFilter;
     GqlTokenMutationResult: GqlTokenMutationResult;
     GqlTokenPrice: GqlTokenPrice;
     GqlTokenPriceChartDataItem: GqlTokenPriceChartDataItem;
@@ -3056,6 +3098,7 @@ export type ResolversParentTypes = ResolversObject<{
     GqlVotingPool: GqlVotingPool;
     Hook: Hook;
     HookData: HookData;
+    HookReviewData: HookReviewData;
     ID: Scalars['ID'];
     Int: Scalars['Int'];
     JSON: Scalars['JSON'];
@@ -3088,6 +3131,16 @@ export interface BytesScalarConfig extends GraphQLScalarTypeConfig<ResolversType
 export interface DateScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['Date'], any> {
     name: 'Date';
 }
+
+export type Erc4626ReviewDataResolvers<
+    ContextType = ResolverContext,
+    ParentType extends ResolversParentTypes['Erc4626ReviewData'] = ResolversParentTypes['Erc4626ReviewData'],
+> = ResolversObject<{
+    reviewFile?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    summary?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    warnings?: Resolver<Array<ResolversTypes['String']>, ParentType, ContextType>;
+    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+}>;
 
 export type GqlBalancePoolAprItemResolvers<
     ContextType = ResolverContext,
@@ -3901,6 +3954,7 @@ export type GqlPoolStableResolvers<
     address?: Resolver<ResolversTypes['Bytes'], ParentType, ContextType>;
     allTokens?: Resolver<Array<ResolversTypes['GqlPoolTokenExpanded']>, ParentType, ContextType>;
     amp?: Resolver<ResolversTypes['BigInt'], ParentType, ContextType>;
+    bptPriceRate?: Resolver<ResolversTypes['BigDecimal'], ParentType, ContextType>;
     categories?: Resolver<Maybe<Array<Maybe<ResolversTypes['GqlPoolFilterCategory']>>>, ParentType, ContextType>;
     chain?: Resolver<ResolversTypes['GqlChain'], ParentType, ContextType>;
     createTime?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
@@ -4182,6 +4236,7 @@ export type GqlPoolTokenDetailResolvers<
     balance?: Resolver<ResolversTypes['BigDecimal'], ParentType, ContextType>;
     balanceUSD?: Resolver<ResolversTypes['BigDecimal'], ParentType, ContextType>;
     decimals?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
+    erc4626ReviewData?: Resolver<Maybe<ResolversTypes['Erc4626ReviewData']>, ParentType, ContextType>;
     hasNestedPool?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
     id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
     index?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
@@ -4639,6 +4694,7 @@ export type GqlTokenResolvers<
     decimals?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
     description?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
     discordUrl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+    erc4626ReviewData?: Resolver<Maybe<ResolversTypes['Erc4626ReviewData']>, ParentType, ContextType>;
     isErc4626?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
     logoURI?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
     name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
@@ -4649,6 +4705,7 @@ export type GqlTokenResolvers<
     telegramUrl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
     tradable?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
     twitterUsername?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+    underlyingTokenAddress?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
     websiteUrl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
     __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 }>;
@@ -4851,6 +4908,7 @@ export type HookResolvers<
     dynamicData?: Resolver<Maybe<ResolversTypes['HookData']>, ParentType, ContextType>;
     enableHookAdjustedAmounts?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
     poolsIds?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
+    reviewData?: Resolver<Maybe<ResolversTypes['HookReviewData']>, ParentType, ContextType>;
     shouldCallAfterAddLiquidity?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
     shouldCallAfterInitialize?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
     shouldCallAfterRemoveLiquidity?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
@@ -4870,6 +4928,17 @@ export type HookDataResolvers<
     addLiquidityFeePercentage?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
     removeLiquidityFeePercentage?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
     swapFeePercentage?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+}>;
+
+export type HookReviewDataResolvers<
+    ContextType = ResolverContext,
+    ParentType extends ResolversParentTypes['HookReviewData'] = ResolversParentTypes['HookReviewData'],
+> = ResolversObject<{
+    name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    reviewFile?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    summary?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+    warnings?: Resolver<Array<ResolversTypes['String']>, ParentType, ContextType>;
     __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 }>;
 
@@ -5282,6 +5351,7 @@ export type Resolvers<ContextType = ResolverContext> = ResolversObject<{
     BigInt?: GraphQLScalarType;
     Bytes?: GraphQLScalarType;
     Date?: GraphQLScalarType;
+    Erc4626ReviewData?: Erc4626ReviewDataResolvers<ContextType>;
     GqlBalancePoolAprItem?: GqlBalancePoolAprItemResolvers<ContextType>;
     GqlBalancePoolAprSubItem?: GqlBalancePoolAprSubItemResolvers<ContextType>;
     GqlBigNumber?: GraphQLScalarType;
@@ -5389,6 +5459,7 @@ export type Resolvers<ContextType = ResolverContext> = ResolversObject<{
     GqlVotingPool?: GqlVotingPoolResolvers<ContextType>;
     Hook?: HookResolvers<ContextType>;
     HookData?: HookDataResolvers<ContextType>;
+    HookReviewData?: HookReviewDataResolvers<ContextType>;
     JSON?: GraphQLScalarType;
     LiquidityManagement?: LiquidityManagementResolvers<ContextType>;
     Mutation?: MutationResolvers<ContextType>;